Rework check_snapshot_mount() v2
This commit tries to make check_snapshot_mount() easier.
The motivation behind the change is it was used incorrectly
from do_delete_snapshot().

In do_delete_snapshot() we only want to check one snapshot
and we don't care about the parent one. So, let's simplify
it to only check one snapshot.

In ploop_merge_snapshot_by_guid() we want to check
two snapshots -- the parent one and the child one. So,
let's just call new check_snapshot_mount() twice -- for
the parent and for the child.

While at it:
1. Change the check_snapshot_mount() to avoid excessive code nesting
2. Move it (and get_devs_str()) to ploop.c as it is used from different places.
3. Improve the error message about busy snapshot -- add snapshot UUID.

[v2: minor code readability improvements]

Signed-off-by: Kir Kolyshkin <kir@openvz.org>