check_snapshot_mount(): add is_device_inuse() check
Temporary snapshots are supposed to be removed as long as they are
not in use anymore. There is a way to mark the temporary snapshot
as being used -- open its corresponding ploop device (see holder_fd
parameter to ploop_create_temporary_snapshot()). Other places dealing
with temporary snapshots check that the snapshot is not in use (see
merge_temporary_snapshots() that calls is_snapshot_in_use()).

Make sure we check it here, too.

[v2: make is_device_inuse() available]

Signed-off-by: Kir Kolyshkin <kir@openvz.org>