normalize_image_name(): fix for slash-ended basedir
After commit f11090d, when basedir already ends with '/' character,
this function stops working correctly.

Ideally, we have to check if basedir ends with slash, add it if missing
then compare the beginning of image file name with that string.
Practically, that would mean copying the basedir string over to
some new place, adding slash, etc.

So, we'd better compare the strings without the slash, and then
compare the last character with the slash.

Bug revealed while testing the previous commit.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>