ploop_merge_snapshot_by_guid(): fix messages
While commit c46f388 did the right thing, the problem with it is
the variable parent_guid used might already be free()'d by
ploop_di_merge_image() above, so we pass a junk pointer to printf.

One solution would be to strdup() parent_guid so it can be used later
in the code. But let's be simple and not print any guid at all, since
we already printed it earlier. What we want at the end is just a
success/failure message.

Also:

 1 print if a snapshot to be merged is temporary just before the merge;

 2 NULLify parent_guid and child_guid to make it clear they are not
   to be used later in the code;

 3 as we print the generic success message, let's print the generic
   failure message, too, in case of failure.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>