make_fs(): retry mkfs without lazy_itable_init
As pointed out in #OVZ-6539, mke2fs doesn't know about
-Elazy_itable_init, so we'd better retry without it if
mkfs fails for the first time.

Also, do the trick and not show stderr from mkfs if running
it for the first time, hiding the potential error about the
unknown lazy_itable_init. When retrying, show stderr, so if
there is any error message printed it will be shown during
the second run.

https://bugs.openvz.org/browse/OVZ-6539

Signed-off-by: Kir Kolyshkin <kir@openvz.org>