make_fs(): simplify mkfs retry logic
Commit 1c6e10a introduced a new flag for run_prg_rc() to hide an error
about non-zero exit code. The same effect can be easily achieved b
asking for the program exit code (i.e. if the last argument
of run_prg_rc() is not NULL).

So, this is a revert of 1c6e10a and an appropriate change in logic.

This is a followup to commit d496ba8.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>