defrag_complete(): check for cmdline
Instead of checking if the running process executable file on disk
is the same as the one that we ran, it's slightly better to check
that the executable path is the same. This is because the file might
be different, for example after defrag rpm upgrade, but its name
and path stays the same.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>