ploop_mount_fs: don't mount ro when finding balloon inode
When mounting a ploop-backed filesystem, we did the following:

1. Mount with MS_RDONLY flag.
2. stat() the balloon file to figure out its inode number
3. Remount with intended flags (usually read-write) plus the
   added balloon_ino=XXX option.

Unfortunately, this way is not working well with applications
that live in MS_SLAVE mount namespace. A slave mount namespace
is a namespace to which all the parent mounts are propagated,
but its own mounts are kept private (i.e. are not propagated
to the parent). In other words, quoting mount(8): "A slave mount
receives propagation from its master, but not vice versa".

Now, I am not sure if this is a bug or a feature, but in case
of a remount changing (adding or removing) MS_RDONLY flag, the
flag is not being properly propagated to the slave mount. So,
when ploop library mounts a filesystem as read-only, and then
remounts it as read-write, a slave mount still have it as read-only:

> # grep ploop /proc/self/mountinfo /proc/41562/mountinfo
> /proc/self/mountinfo:114 141 182:467617 / /pcs/mnt/vol33 rw,relatime shared:65 - ext4 /dev/ploop29226p1 rw,data=ordered,balloon_ino=12
> /proc/41562/mountinfo:115 155 182:467617 / /pcs/mnt/vol33 ro,relatime master:65 - ext4 /dev/ploop29226p1 rw,data=ordered,balloon_ino=12
>                                                           ^^

Here PID 41562 lives in slave mount namespace. As you can see, its
mountingo still shows "ro" in mount options (field 6), despite
the fact that it has been remounted read-write (as shown in
/proc/self/mountinfo which is in initial mount namespace).

This slave mount pecularity affects Docker on a RHEL7 system,
which is set up so that docker daemon runs in slave mount namespace
(with the help from systemd, by having "MountFlags=slave" line in
/usr/lib/systemd/system/docker.service). So, when we start docker
daemon, then mount a ploop, then try to use it as a volume for
container ("docker run -v /path/on/host:/path/in/container ..."),
it appears as read-only inside the container:

 # ploop list -a
 ploop29226   /pcs/img/vol33/root.hdd /pcs/mnt/vol33
 [root@kube-node1 ploop]# touch /pcs/mnt/vol33/dude
 # docker run -it -v /pcs/mnt/vol33:/noway:rw centos sh
 sh-4.2# touch /noway/dude
 touch: cannot touch '/noway/dude': Read-only file system

Initially mounting ploop-backed fs as read-write (more specifically,
with the same flags as intended) fixes this glitch, and currently
I am not able to foresee any bad implications of doing it.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>
Cc: Maxim Patlasov <mpatlasov@openvz.org>
Cc: Andrey Vagin <avagin@openvz.org>