defrag_complete(): don't use %ms in fscanf
Unfortunately, on a RHEL5 system (which we still support!) we have
pretty ancient gcc-4.1.2 and glibc-2.5, and they don't support %ms
in fscanf, spitting the following error:

cc1: warnings being treated as errors
balloon.c: In function 'defrag_complete':
balloon.c:1210: warning: unknown conversion type character 'm' in format
balloon.c:1210: warning: too many arguments for format

In fact, it's more like an error rather than a warning, because cmd
wouldn't be allocated and read into. We could have used %as instead
(it's what it was called back in the day) but this won't work with
a modern gcc/glibc.

So, just use read_line with a long enough buffer. This also makes the
code simpler and adds nice error reporting in case the file can't be
read.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>